+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * 3D geometry support, waypoint geometries may no longer specify elevation values
   [#150](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/150)
+* Support for comments in route waypoints
+  [#141](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/141)
 
 ### Added
 
@@ -26,6 +28,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   [#82](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/82)
 * improved Continuous Integration, verifying test network can be minimally processed
   [#168](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/168)
+* versioning to the test network
+  [#173](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/173)
+* script to recreate test network as GeoJSON for testing in QGIS
+  [#174](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/174)
 
 ### Changed
 
@@ -33,6 +39,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   [#140](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/140)
 * downgrading required Python version to 3.8, for compatibility with the Operations Data Store project
   [#138](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/138)
+* incorporating the test network into this project
+  [#172](https://gitlab.data.bas.ac.uk/MAGIC/air-unit-network-dataset/-/issues/172)
 
 ## [0.2.2] - 2022-10-21